Fixed compilation on Windows with Visual Studio
git-svn-id: file:///home/michael/svn/yuni/trunk@2028 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/logs/decorators/message.h
@@ -26,7 +26,7 @@ namespace Logs
 				System::Console::TextColor<VerbosityType::messageColor>::Set(out);
 
 			// The message
-			out.write(s.c_str(), (size_t)s.sizeInBytes());
+			out.write(s.c_str(), (unsigned int)s.sizeInBytes());
 
 			// Resetting the color
 			if (VerbosityType::messageColor != System::Console::none && Handler::unixColorsAllowed)
